firmware:  - PSWF changes.

+++ firmware/app/mrd/voiceserviceinterface.cpp
@@ -131,7 +131,7 @@ void VoiceServiceInterface::setCurrentChannel(ChannelStatus status) {
     currentChannelChanged();
 }
 
-void VoiceServiceInterface::fistPacketRecieve(uint8_t packet)
+void VoiceServiceInterface::fistPacketRecieve(int packet)
 {
    firstPacket(packet);
 }
